# public/docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "version": "1.0.0",
    "title": "Swagger MyProfile",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "demo.intellifi.tech/demo/MyProfile/web/public",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/event/search": {
      "post": {
        "tags": [
          "event"
        ],
        "summary": "Etkinlik arama",
        "description": "Gönderilen title keyine göre arama yapar.",
        "operationId": "searchEvent",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/EventSearch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Events"
            }
          },
          "204": {
            "description": "No content"
          }
        }
      }
    },
...
    "/message/getMessages": {
      "post": {
        "tags": [
          "message"
        ],
        "summary": "Kullanıcının tüm mesajlarını getirir.",
        "description": "This can only be done by the logged in user.",
        "operationId": "getMessages",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/ApiToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    },
...
    "/message/answerMessage": {
      "post": {
        "tags": [
          "message"
        ],
        "summary": "Mesaj yanıtla",
        "description": "This can only be done by the logged in user.",
        "operationId": "AnswerMessage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/AnswerMessage"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "ApiToken": {
      "type": "object",
      "properties": {
        "api-token": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
    "Events": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "bigInteger"
        },
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "activity_date": {
          "type": "string"
        },
        "end_date": {
          "type": "string"
        },
        "latitude": {
          "type": "string"
        },
        "longitude": {
          "type": "string"
        },
        "comments": {
          "$ref": "#/definitions/Comments"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "EventSearch": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "meterLimit": {
          "type": "string"
        },
        "latitude": {
          "type": "string"
        },
        "longitude": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
    "Comments": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "event_id": {
          "type": "integer"
        },
        "comment": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
    "Message": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "type": {
          "type": "integer"
        },
        "parent_id": {
          "type": "integer"
        },
        "from_user_id": {
          "type": "integer"
        },
        "to_user_id": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
    "AnswerMessage": {
      "type": "object",
      "properties": {
        "api_token": {
          "type": "string"
        },
        "parent_id": {
          "type": "integer"
        },
        "to_user_id": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        }
      },
      "xml": {
        "name": "User"
      }
    },
...
  },
...
}